Update base to Docker 19
It looks like the drone config and manifest support tags, so it would
probably be good to tag the current master as 18 and then, after this is
merged, tag it as 19.0 or something so that it can get a 19 tag.

If there is a desire to preserve minor or patch Docker versions, a more
complex build and tagging system could be set up. I figure at least
getting a major version update would be helpful though.